fix: code fixed

+++ src/main/java/lang/print/gaps/finalModuleTask/TemperatureConverter.java
@@ -3,7 +3,7 @@ package lang.print.gaps.finalModuleTask;
 public class TemperatureConverter {
     public void toFahrenheit(int temperatureCelsius){
         float result;
-        result = temperatureCelsius * 9 / 5 + 32;
+        result = (float)temperatureCelsius * 9 / 5 + 32;
         System.out.println(result);
     }
 }
